In this commit: -Create a comments popup only with the top section -When the popoup loads, the webapp retrieves data from the pokeapi -Shows details about the selected pokemon -Add event listener to the comments buttons -Resolve merge issues with development
