+++ src/scripts/modules/generators.js
@@ -31,7 +31,6 @@ export default async () => {
 
     const buttonHeart = document.createElement('button');
     buttonHeart.classList.add('heart');
-    buttonHeart.id = id;
     const heart = new Image();
     heart.src = Heart;
     buttonHeart.appendChild(heart);
@@ -40,6 +39,7 @@ export default async () => {
     card.appendChild(info);
 
     const commentsButton = document.createElement('button');
+    commentsButton.id = id;
     commentsButton.textContent = 'Comments';
     card.appendChild(commentsButton);
   });
