--- src/scripts/popup.js
@@ -1,5 +1,0 @@
-export const getPokemonInfo = async (id) => {
-  const getInfo = await fetch(`https://pokeapi.co/api/v2/pokemon/${id}/`);
-  const json = await getInfo.json();
-  return json;
-};
