+++ src/scripts/styles/pages/_home.scss
@@ -3,49 +3,65 @@
   position: fixed;
   height: 100vh;
   width: 100vw;
-  background-color: rgba(255, 255, 255, 0.774);
-  display: flex;
+  background-color: rgba(255, 255, 255, 0.9);
+  display: none;
   flex-direction: column;
   justify-content: center;
   align-items: center;
   gap: 2%;
+
   &-close {
     position: absolute;
     top: 2%;
     right: 2%;
     height: 50px;
     width: 50px;
+
     img {
       width: 100%;
       height: 100%;
     }
+
+    &:hover {
+      cursor: pointer;
+    }
   }
+
   &-img {
     height: 300px;
     width: 300px;
     border: 2px solid black;
     border-radius: 5%;
     padding: 10px;
+
     img {
       width: 100%;
       height: 100%;
     }
   }
+
   h2 {
     font-size: 3rem;
   }
+
   &-info {
     width: 320px;
     display: flex;
     flex-wrap: wrap;
     justify-content: space-between;
+
     p {
       width: 155px;
       margin: 10px 0;
       font-size: 1.5rem;
+
       &:nth-child(2n) {
         text-align: right;
       }
+    }
+
+    :last-child {
+      width: 100%;
     }
   }
 }
