+++ src/template.html
@@ -36,6 +36,7 @@
       <p class="abilities"></p>
       <p class="height"></p>
       <p class="weight"></p>
+      <p class="moves"></p>
     </div>
   </section>
 </body>
